requirements.txt
pillow
qrcode[pil]
reportlab